requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0.post2
requests==2.31.0
aiofiles==23.2.1
websockets==11.0.3
httpx==0.24.1
psycopg2-binary==2.9.7
alembic==1.12.0 